Small updates for Task8
Changed task 8 so it can be called and ran from testing.py or similar (only requires argument of k)

Note: path_to_imgs and path_to_metadata currently hardcoded. I also adjusted the metadata file to be able to use same format as the metadata file which we have created and used for testing.

This will need to be decided upon.